@@ -1,22 +1,21 @@
-import { BrowserRouter, Route, Switch } from "react-router-dom";
 import "./App.css";
-import Header from "./components/Header/Header";
-import SimpleBottomNavigation from "./components/MainNav";
-import Movies from "./Pages/Movies/Movies";
-import Series from "./Pages/Series/Series";
-import Trending from "./Pages/Trending/Trending";
-import Search from "./Pages/Search/Search";
+import { BrowserRouter, Route, Switch } from "react-router-dom";
+import Header from "./component/Header/Header";
 import { Container } from "@material-ui/core";
-
+import Trending from "./Page/Trending/Trending";
+import Movie from "./Page/Movie/Movie";
+import Series from "./Page/Series/Series";
+import Search from "./Page/Search/Search";
+import SimpleBottomNavigation from "./component/MainNavbar";
 function App() {
   return (
     <BrowserRouter>
       <Header />
-      <div className="app">
+      <div className="App">
         <Container>
           <Switch>
             <Route path="/" component={Trending} exact />
-            <Route path="/movies" component={Movies} />
+            <Route path="/movie" component={Movie} />
             <Route path="/series" component={Series} />
             <Route path="/search" component={Search} />
           </Switch>
